Fix -Wconversion warnings, phase 2
empire.c

This one was rife with potential integer overflow if the input
variables were large enough. I've just increased the size of the
types used until everything would fit, so there's some int64_t in
the API now.

+++ src/doors/clans-src/src/empire.h
@@ -17,9 +17,9 @@ void DonateToEmpire(struct empire *Empire);
 
 
 int16_t ArmySpeed(struct Army *Army);
-int32_t ArmyOffense(struct Army *Army);
-int32_t ArmyDefense(struct Army *Army);
-int32_t ArmyVitality(struct Army *Army);
+int64_t ArmyOffense(struct Army *Army);
+int64_t ArmyDefense(struct Army *Army);
+int64_t ArmyVitality(struct Army *Army);
 
 void ProcessAttackPacket(struct AttackPacket *AttackPacket);
 void ProcessResultPacket(struct AttackResult *Result);
